backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1
loguru==0.7.2
jinja2==3.1.4
orjson==3.10.7
uvloop==0.20.0
httptools==0.6.1
httpx==0.27.0

# NLP
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.1

# Parsing
pdfminer.six==20240706
python-docx==1.1.2
beautifulsoup4==4.12.3
chardet==5.2.0

# Email
imapclient==3.0.1
click==8.1.7
